Move timeline folder to workspace view folder
