+++ vue/src/core/components/layout/DefaultLayout/DefaultLayout.vue
@@ -25,8 +25,8 @@ import {mapState, createNamespacedHelpers, mapActions} from 'vuex';
 import MenuService from "../../sidebar/MenuService";
 import MenuItem from "../../sidebar/MenuItem";
 import EmployeeFormModal from "@/modules/setup/employees/EmployeeFormModal";
-import TimelineForm from "@/modules/Timeline/TimelineForm";
-import TimelineShare from "@/modules/Timeline/TimelineShare";
+import TimelineForm from "@/modules/Workspace/view/timeline/TimelineForm";
+import TimelineShare from "@/modules/Workspace/view/timeline/TimelineShare";
 
 const {mapActions: userMapActions} = createNamespacedHelpers('user');
 const {mapState: mapStateWorkspace, mapActions: mapActionsWorkspace} = createNamespacedHelpers('workspace');
